Add FacialExpressionDisable and change local smoothing to 0.7 for all binary
FacialExpressionDisable bool can be used in disabling gesture animations without using face tracking bools to allow toggling when desired instead of hard block. Change the binary smoothing to 0.7 for local as that how the old setup worked with smoothing the lower values were causing stepping to be seen for the binary parameters. Floats will remain at existing. Updated menu textures to be more pleasing to the eye.