Limit Eye Rotation
Typical values for eye rotation max at 0.7, Meta release update on v54 firmware that cause saturation to 1 even though the OpenXR module compensated to be from 0 to 0.7. It is like exponential increase in rotation after 0.7. This is not useful and looks bad. So I am adding a limit on the rotation in the template.