VRCFury Tracking Control Logic Change
VRCFury Tracking Control Logic Change

VRCFury recent made change that merges all the tracking states in the avatar. This is good directions as state controls are hard to manage if they are on different layers. The logic used does not work correctly with the merging with VRCFury so logic was changed to be compatible.

Bump Version